backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.22
google-generativeai==0.5.4
reportlab==4.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0
gunicorn==22.0.0
